Fix for MDRV2 GetDataInformation attribute
Issue: Incorrect datatype of dataVersion and format of data response
for GetDataInformation D-Bus method call. So getting incorrect response
in MDRV2 GetDataInformation API.

Fix: Corrected datatype of dataVersion and  code to properly convert
uint32_t to byte vector.

Tested:
Before Fix:
Command:  busctl call xyz.openbmc_project.Smbios.MDR_V2
          /xyz/openbmc_project/Smbios/MDR_V2
          xyz.openbmc_project.Smbios.MDR_V2 GetDataInformation y 0
Response: ay 21 2 40 41 42 43 44 45 46 47 48 49 50 51 52 53 54 66 1 69
          15 27
After Fix:
Command:  busctl call xyz.openbmc_project.Smbios.MDR_V2
          /xyz/openbmc_project/Smbios/MDR_V2
          xyz.openbmc_project.Smbios.MDR_V2 GetDataInformation y 0
Response: ay 27 2 40 41 42 43 44 45 46 47 48 49 50 51 52 53 54 66 1 0 0
          17 69 15 0 1 64 219

+++ include/smbios_mdrv2.hpp
@@ -84,7 +84,7 @@ typedef struct
     DataIdStruct id;
     uint32_t size;
     uint32_t dataSetSize;
-    uint32_t dataVersion;
+    uint8_t dataVersion;
     uint32_t timestamp;
 } Mdr2DirEntry;
